pyproject.toml
[project]
name = "triage-poc"
version = "0.1.0"
requires-python = "==3.12.*"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "python-frontmatter>=1.1.0",
    "structlog>=24.2.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
